requirements.txt
aniso8601==9.0.1
asttokens==2.2.1
attrs==22.2.0
aubio==0.4.9
backcall==0.2.0
brotlipy==0.7.0
certifi
cffi
charset-normalizer
clams-python==0.5.3
click==8.1.3
comm==0.1.2
cryptography
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.3.0
executing==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==2.2.3
Flask-RESTful==0.3.9
future==0.18.3
gunicorn==20.1.0
idna
imageio==2.25.1
ipykernel==6.21.2
ipython==8.10.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.2.0
lapps==0.0.2
lxml==4.9.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mkl-service==2.4.0
mmif-python==0.4.8
nest-asyncio==1.5.6
nltk==3.7
numpy==1.24.2
ordered-set==4.1.0
packaging==23.0
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.0.0
pluggy 
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycosat
pycparser
pydantic==1.10.7
pydub==0.25.1
pyglet==1.5.26
pyOpenSSL
pyrsistent==0.19.3
PySocks
python-dateutil==2.8.2
pytz==2022.7.1
PyWavefront==1.3.3
pyzmq==25.0.0
regex==2022.6.2
requests
requests-file==1.5.1
requests-toolbelt==0.10.1
ruamel.yaml
ruamel.yaml.clib
six
stack-data==0.6.2
toolz
tornado==6.2
tqdm==4.64.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3
wcwidth==0.2.6
Werkzeug==2.2.3
zeep==4.2.1
zstandard
